// src/app/models/dt/dtasist.ts
export class dtapert{    
    equ_cont: number;
    equ_codi: string;
    equ_nomb: string;
}

export class dtasist{
    constructor(){}
    emp_codi: number;
    ter_coda: string;
    equ_cont: number;
    asi_desc: string;
}

export class dtasise{
    emp_codi: number;
    asi_cont: number;
    deq_iden: string;
    deq_noco: string;
    das_asis: boolean;
    das_cont: number;
}